requirements.txt
langchain==0.0.292
python-dotenv==1.0.0
streamlit==1.26.0
openai==0.28.0